RIOT-129 and RIOT-128: Persistence -> Added methods to get the total number of items / filtered items; REST -> Added infos (offset, limit, total) at response header for get requests using pagination and/or filtering; Extended Unit Tests; Updated Documentation.

+++ commons/src/main/java/de/uni_stuttgart/riot/commons/rest/data/FilteredRequest.java
@@ -14,8 +14,8 @@ public class FilteredRequest {
 
     private boolean orMode;
     private List<FilterAttribute> filterAttributes;
-    private long offset = 0;
-    private long limit = 0;
+    private int offset = 0;
+    private int limit = 0;
 
     public boolean isOrMode() {
         return orMode;
@@ -33,19 +33,19 @@ public class FilteredRequest {
         this.filterAttributes = filterAttributes;
     }
 
-    public long getOffset() {
+    public int getOffset() {
         return offset;
     }
 
-    public void setOffset(long offset) {
+    public void setOffset(int offset) {
         this.offset = offset;
     }
 
-    public long getLimit() {
+    public int getLimit() {
         return limit;
     }
 
-    public void setLimit(long limit) {
+    public void setLimit(int limit) {
         this.limit = limit;
     }
 }
